Initial: def, declare, defn, doseq. Reword if, let. Clarification attempt.
Reworded 'if' to be more clear about truthiness.

Added clarifcation to let to indicate prefernce for creating local
bindings. Also present in def. This will likely need to be more
carefully worded in def with reference to var usage.

Started refering to function/macros in documentation with (name ...) to
allow for more visually distinct reference as opposed to the literal
word.