Document java.jdbc 0.5.5 changes
In version 0.5.5, the unrolled keyword/value arguments are
deprecated in favor of a hash map of options as a single argument (in
order to make the functions easier to compose with other functions). In
version 0.6.0, support for the unrolled keyword/value arguments will be
removed (so I'll update the documentation again when that happens).